fix: cast float to int for jlimit template
Compiler error on Mac: jlimit deduced conflicting types for int/float.
Adding explicit static_cast<int> to fix.

+++ Source/PedalComponent.cpp
@@ -241,8 +241,8 @@ void PedalComponent::paint(juce::Graphics& g)
         if (skin.isValid())
         {
             auto resizedSkin = skin.rescaled(
-                juce::jlimit(1, 1000, face.getWidth()),
-                juce::jlimit(1, 1000, face.getHeight()),
+                juce::jlimit(1, 1000, static_cast<int>(face.getWidth())),
+                juce::jlimit(1, 1000, static_cast<int>(face.getHeight())),
                 juce::Graphics::highResamplingQuality);
             g.drawImage(resizedSkin, face.toNearestInt());
         }
